# dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Bice",
        "description": "microservicio de valores",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "{protocol}://localhost:5000/",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                }
            }
        },
        {
            "url": "{protocol}://bice-backent-prueba.herokuapp.com/",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "https"
                }
            }
        }
    ],
    "components": {
        "schemas": {
            "ResponseProducto": {
                "type": "object",
                "required": [
                    "ok"
                ],
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "datos": {
                        "type": "object",
                        "properties": {
                            "key": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "unit": {
                                "type": "string"
                            },
                            "date": {
                                "type": "string"
                            },
                            "value": {
                                "type": "number"
                            }
                        }
                    }
                }
            },
            "ResponseTipo": {
                "type": "object",
                "required": [
                    "ok"
                ],
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "datos": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "cobre"
                        }
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
        "/valores/": {
            "post": {
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "fecha": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Valores"
                ],
                "summary": "ver productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseProducto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "NOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tipos/": {
            "get": {
                "tags": [
                    "Valores"
                ],
                "summary": "ver productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseTipo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "NOK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}